requirements.txt
PyQt5==5.14.1
Pillow==7.0.0